% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/average_tax_table.R
\name{average_tax_table}
\alias{average_tax_table}
\title{Create an average tax table comparing two scenarios using the `calculate_tax` function}
\usage{
average_tax_table(inc_start = 1000, inc_stop = 3e+05, step = 1000,
  ...)
}
\arguments{
\item{inc_start}{Income range start}

\item{inc_stop}{Income range stop}

\item{step}{Increment amount}

\item{...}{Paramaters for the `calculate_tax` function}
}
\value{
A data frame showing average tax rates by income level for two scenarios
}
\description{
Create an average tax table comparing two scenarios using the `calculate_tax` function
}
\examples{
average_tax_table(base_tax_brackets = c(18200, 37000, 80000, 1.8e5, Inf),
                  change_tax_brackets = c(18200, 37000, 87000, 1.8e5, Inf))

}
