% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createArtificalSociety.R
\name{createArtificalSociety}
\alias{createArtificalSociety}
\title{Creates a homophilous Social Influence Network}
\usage{
createArtificalSociety(society = society, homophily = homophily, nu = 4.5)
}
\arguments{
\item{society}{tibble containing society}

\item{homophily}{tibble with homophily parameters for society}

\item{nu}{gower distance exponent}
}
\value{

}
\description{
Creates a homophilous Social Influence Network
}
