% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distro_kable.R
\name{distro_kable}
\alias{distro_kable}
\title{Pretty table for distributional analysis}
\usage{
distro_kable(x)
}
\arguments{
\item{x}{A microsim object}
}
\description{
Pretty table for distributional analysis
}
