% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/function_difference.R
\name{function_difference}
\alias{function_difference}
\title{Compare tax owed or benefit provided from tax/transfer calculators}
\usage{
function_difference(income, FUN, base_arguments, change_arguments)
}
\arguments{
\item{income}{Taxable income}

\item{FUN}{Function to calculate}

\item{base_arguments}{Arguments under the base scenario}

\item{change_arguments}{Arguments under the change scenario}
}
\value{
A list with three elements, the tax owed/benefit provided under the
  base and change scenarios and the difference between the two scenarios
}
\description{
Compare tax owed or benefit provided from tax/transfer calculators
}
\examples{
function_difference(100000, income_tax,
base_arguments = list(brackets = c(18200, 37000, 87000, 180000, Inf)),
change_arguments = list(brackets = c(18200, 37000, 90000, 180000, Inf)))
}
