% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/microsim.R
\name{microsim}
\alias{microsim}
\title{Perform a microsimulation of tax changes using the 2013-14 two per cent sample of Australian taxpayers}
\usage{
microsim(keep_df = FALSE, ...)
}
\arguments{
\item{keep_df}{Whether to keep the amended tax file, mainly useful for debugging.}

\item{...}{Parameters for the `calculate_tax` function}
}
\value{
A microsim object
}
\description{
Perform a microsimulation of tax changes using the 2013-14 two per cent sample of Australian taxpayers
}
\examples{
microsim(base_tax_brackets = c(18200, 37000, 80000, 1.8e5, Inf),
         change_tax_brackets = c(18200, 37000, 87000, 1.8e5, Inf))

}
