% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.microsim.R
\name{plot.microsim}
\alias{plot.microsim}
\title{A plot method for microsim objects}
\usage{
\method{plot}{microsim}(x, type = "average", base_title = "base",
  change_title = "change")
}
\arguments{
\item{x}{Microsimulation to plot}

\item{type}{Whether to show a plot of average tax rates or a distributional plot}
}
\description{
A plot method for microsim objects
}
