% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.microsim.R
\name{print.microsim}
\alias{print.microsim}
\title{Print method for microsim objects}
\usage{
\method{print}{microsim}(x)
}
\arguments{
\item{x}{A microsim object}
}
\description{
Print method for microsim objects
}
