% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/society-data.R
\docType{data}
\name{society}
\alias{society}
\title{Agent demographic variables.}
\format{
A data frame with 924 rows and 12 variables:
\describe{
\item{ID}{agent ID 1-924}
\item{age}{5 age ranges  "25-34" "35-44" "45-54" "18-24" "55+" }
\item{class}{two classes }
\item{gender}{ }
...
}
}
\source{
2018 Survey
}
\usage{
society
}
\description{
A dataset derives from a subset of 2018 survey variables for drivers or prospective drivers
}
\keyword{datasets}
