% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.microsim.R
\name{summary.microsim}
\alias{summary.microsim}
\title{Summary method for microsim objects}
\usage{
\method{summary}{microsim}(x)
}
\arguments{
\item{x}{A microsim object}
}
\description{
Summary method for microsim objects
}
