% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uprate_data.R
\name{uprate_data}
\alias{uprate_data}
\title{Uprade data by wage price index}
\usage{
uprate_data(tax_file)
}
\arguments{
\item{tax_file}{}
}
\value{
An uprated tax sample
}
\description{
Uprade data by wage price index
}
