% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/variable_compound.R
\name{variable_compound}
\alias{variable_compound}
\title{Compound a principal by varying rates of interest.}
\usage{
variable_compound(p, rates)
}
\arguments{
\item{p}{Principal}

\item{rates}{Vector of interest rates for each period}
}
\value{
The compounded principal
}
\description{
Compound a principal by varying rates of interest.
}
\examples{
variable_compound(100, c(.03, .03, .04))
}
