% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/microsim.R
\name{microsim}
\alias{microsim}
\title{Perform a microsimulation of tax changes using the 2016-17 two per cent
sample of Australian taxpayers.}
\usage{
microsim(keep_df = FALSE, employment_growth = c(275, 0.015, 0.015,
  0.0125), wages_growth = c(0.0225, 0.0275, 0.0325, 0.035),
  parallel = FALSE, ...)
}
\arguments{
\item{keep_df}{Whether to keep the amended tax file, mainly useful for
debugging.}

\item{employment_growth}{A vector containing forecasts for employment growth}

\item{wages_growth}{A vector containing forecasts for wages growth}

\item{parallel}{Should this compute using the parallel package}

\item{...}{Parameters for the `calculate_tax` function}
}
\value{
A microsim object
}
\description{
The analysis year is 2019-20. The assumptions for employment growth and wages
are from the 2018-19 Budget and are as follows.

Wages 2.25%, 2.75%, 3.25%, 3.5%
Employment 2.75%, 1.5%, 1.5%, 1.25%
}
\examples{
microsim(base_tax_brackets = c(18200, 37000, 80000, 1.8e5, Inf),
         change_tax_brackets = c(18200, 37000, 87000, 1.8e5, Inf))

}
