% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uprate_data.R
\name{uprate_data}
\alias{uprate_data}
\title{Uprade income data by vector of rates}
\usage{
uprate_data(tax_file, uprate_vector)
}
\arguments{
\item{tax_file}{}
}
\value{
An uprated tax sample
}
\description{
Uprade income data by vector of rates
}
